ADD: add utils.py to separate run script and some tools
